feat: add PostPoll and PostLottery components
